+++ riscv/insns/fcvt_s_w.h
@@ -1,4 +1,4 @@
 require_fp;
 softfloat_roundingMode = RM;
-FRD = i32_to_f32(RS1);
+FRD = i32_to_f32((int32_t)RS1);
 set_fp_exceptions;
